test folder clean
