+++ cypress/e2e/pages/applications/edit/[id]/person/edit-household-member.cy.ts
@@ -55,4 +55,15 @@ describe('Edit household member', () => {
     ApplicationEditHouseholdMember.getErrorSummary().should('be.visible');
     cy.contains(`Unable to update application (${errorCode})`);
   });
+  it('shows access denied page for user with read only permissions', () => {
+    const applicationId = faker.string.uuid();
+    const personId = faker.string.uuid();
+
+    cy.clearAllCookies();
+    cy.loginAsUser('readOnly');
+    cy.visit(
+      `applications/edit/${applicationId}/${personId}/edit-household-member`
+    );
+    cy.contains('Access denied');
+  });
 });
