+++ cypress/e2e/pages/applications/edit/add-household-member.cy.ts
@@ -71,4 +71,12 @@ describe('Add household member', () => {
     cy.contains(`Unable to update application (${errorCode})`);
     AddHouseholdMemberPage.getErrorSummary().should('be.visible');
   });
+  it('shows access denied page for user with read only permissions', () => {
+    const applicationId = faker.string.uuid();
+
+    cy.clearAllCookies();
+    cy.loginAsUser('readOnly');
+    cy.visit(`applications/edit/${applicationId}/add-household-member`);
+    cy.contains('Access denied');
+  });
 });
