+++ cypress/e2e/pages/applications/view-register.cy.ts
