+++ cypress/e2e/pages/apply/submit/ethnicity-questions.cy.ts
