+++ cypress/e2e/pages/apply/verify.cy.ts
@@ -1,10 +1,11 @@
 import { faker } from '@faker-js/faker';
-import VerifyPage from '../../pages/verify';
-import SignInPage from '../../pages/signIn';
-import { generateApplication } from '../../../testUtils/applicationHelper';
-import StartPage from '../../pages/start';
-import { generateEmailAddress } from '../../../testUtils/personHelper';
+import VerifyPage from '../../../pages/verify';
+import SignInPage from '../../../pages/signIn';
+import { generateApplication } from '../../../../testUtils/applicationHelper';
+import StartPage from '../../../pages/start';
+import { generateEmailAddress } from '../../../../testUtils/personHelper';
 import { StatusCodes } from 'http-status-codes';
+import HomePage from '../../../pages/home';
 
 const applicationId = faker.string.uuid();
 const personId = faker.string.uuid();
@@ -87,5 +88,44 @@ describe('verify code page', () => {
 
     //should not push to the next page
     StartPage.getStartApplicationPage().should('not.exist');
+  });
+  it(`user successfully verifies and starts application`, () => {
+    const generateEmailAddress = faker.internet.email({
+      provider: 'hackneyTEST.gov.uk',
+    });
+    const generateCode = faker.number
+      .int({ min: 100000, max: 999999 })
+      .toString();
+    const applicationId = faker.string.uuid();
+    const personId = faker.string.uuid();
+    const application = generateApplication(
+      applicationId,
+      personId,
+      false,
+      false,
+      false
+    );
+
+    cy.clearCookies();
+
+    cy.mockHousingRegisterApiPostGenerateToken(0, true);
+    cy.mockHousingRegisterApiPostVerifyToken(applicationId);
+    cy.mockHousingRegisterApiGetApplications(applicationId, application, true);
+
+    HomePage.visit();
+    HomePage.getCookiesButton().click();
+    HomePage.getStartApplicationButton().scrollIntoView().click();
+
+    SignInPage.getEmailInput().scrollIntoView().type(generateEmailAddress);
+    SignInPage.getSubmitButton().scrollIntoView().click();
+    cy.get('button[type="submit"]').click();
+
+    VerifyPage.getVerifyCodePage().should('be.visible');
+    VerifyPage.getVerifyCodeInput().scrollIntoView().type(generateCode);
+
+    cy.loginAsResident(applicationId, false, true);
+    VerifyPage.getVerifySubmitButton().scrollIntoView().click();
+
+    StartPage.getStartApplicationPage().should('be.visible');
   });
 });
